✨ [BE] Feat: 예금 단건 조회 API 개발

+++ Backend/src/main/java/com/ssafy/teentech/deposit/repository/DepositRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface DepositRepository extends JpaRepository<Deposit,Long> {
+public interface DepositRepository extends JpaRepository<Deposit,Integer> {
     Optional<List<Deposit>> findAllByUser(User user);
 }
